refactor(webui): split channel setup responsibilities
