+++ webui/src/components/settings/system/ChannelsSettings.tsx
@@ -6,9 +6,10 @@ import { Input } from "@/components/ui/input";
 import { SegmentedControl } from "@/components/ui/segmented-control";
 import { SETTINGS_SEARCH_INPUT_CLASS } from "@/components/settings/shared/SettingsControls";
 import type { ChannelFeatureAction } from "@/channel-plugins/types";
+import { ChannelCatalogRow } from "@/components/settings/channels/ChannelCatalogRow";
 import { localizedChannelDisplayName } from "@/components/settings/channels/ChannelIdentity";
 import { ChannelHelpMenu } from "@/components/settings/channels/ChannelHelpMenu";
-import { ChannelCatalogRow, ChannelSetupPanel } from "@/components/settings/channels/ChannelSetupPanel";
+import { ChannelSetupPanel } from "@/components/settings/channels/ChannelSetupPanel";
 import { DismissibleStatusMessage, RestartRequiredNotice, SettingsGroup } from "@/components/settings/shared/SettingsControls";
 import type { NanobotFeaturesPayload } from "@/lib/types";
 import { Dialog, DialogClose, DialogContent, DialogTitle } from "@/components/ui/dialog";
